fix: fixed @imyanice code about attendance
